dreamspace-ai/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Pillow==10.0.1
opencv-python==4.8.1.78
ultralytics==8.0.204
torch==2.1.0
torchvision==0.16.0
diffusers==0.21.4
transformers==4.35.0
accelerate==0.24.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
boto3==1.29.7
numpy==1.24.3
requests==2.31.0
gunicorn==21.2.0